Racer tourney CRUD (tests only)

+++ database/migrations/2021_06_17_091621_create_tourneys_table.php
@@ -14,7 +14,7 @@ class CreateTourneysTable extends Migration
             $table->string('track_id', 4);
             $table->string('room', 20);
             $table->timestamp('started_at');
-            $table->unsignedTinyInteger('signup_time');
+            $table->unsignedSmallInteger('signup_time');
             $table->foreignId('supervisor_id')->nullable()->constrained('users')->cascadeOnUpdate()->nullOnDelete();
             $table->string('supervisor_username', 16);
             $table->string('status', 20)->default('planned');
